bring back virtual base for digital actuators
